Fix etcd support
etcd v3.5.0 (June 2021) no longer emits a "listening for peers on" log
message on startup [1]. The new message is "serving peer traffic". This
is the second time this has happened to us [2]. Rather than chasing this
moving car, wait on the port to come up instead. This is likely good
enough and allows us to continue supporting older (pre-v3.5.0) versions.

[1] https://github.com/etcd-io/etcd/commit/74d5ba577
[2] https://github.com/jb/pifpaf/commit/7496e5e1c

Signed-off-by: Stephen Finucane <stephen@that.guru>